Remove legacy realtime service and update plan
