+++ README.md
@@ -71,3 +71,41 @@ Yes, you can!
 To connect a domain, navigate to Project > Settings > Domains and click Connect Domain.
 
 Read more here: [Setting up a custom domain](https://docs.lovable.dev/tips-tricks/custom-domain#step-by-step-guide)
+
+## Security Logging
+
+This project uses `EnhancedSecurityAuditService` for all audit and security events. Previous services like `auditLogService` and `secureAuditLogService` were removed in favor of this consolidated implementation.
+
+## Development Notes
+
+### Linting
+Run `npm run lint` to check code style. The project uses a minimal ESLint configuration without extra plugins.
+
+## Component Consolidation
+
+Legacy implementations of several major features have been removed. The app now
+uses the enhanced versions exclusively:
+
+- **EnhancedCBTSkillsLibrary** replaces the basic CBT skills library
+- **EnhancedCrisisSystem** replaces the old crisis intervention system
+- **EnhancedCalendarPage** replaces the previous calendar page
+- **dashboard/NotificationBanner** replaces the generic banner component
+- **useSecureAuditLogger** replaces useAuditLogger and server-side variants
+- **EnhancedSecurityAuditService** consolidates audit logging services
+- **EnhancedInputValidator** consolidates input validation utilities
+- **enhancedRealtimeService** now handles all realtime features
+
+These components provide richer functionality and improved security compared to
+their predecessors.
+
+See `docs/DEDUPLICATION_PLAN.md` for details about remaining duplicates and the
+migration checklist.
+
+### Deployment to Vercel
+
+This project can be deployed on [Vercel](https://vercel.com). A `vercel.json` file is included to configure the build command and output directory. Vercel will run `npm run build` and serve the generated static files from the `dist` directory.
+
+1. Import the repository into Vercel.
+2. Ensure required environment variables (e.g. `VITE_SUPABASE_URL`, `VITE_SUPABASE_ANON_KEY`) are set in the project settings.
+3. Trigger a deployment.
+
