+++ src/components/assessments/PHQ2Assessment.tsx
@@ -3,7 +3,7 @@ import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import { Button } from '@/components/ui/button';
 import { RadioGroup, RadioGroupItem } from '@/components/ui/radio-group';
 import { Label } from '@/components/ui/label';
-import { useAuditLogger } from '@/hooks/useAuditLogger';
+import { useSecureAuditLogger } from '@/hooks/useSecureAuditLogger';
 import { escalateCrisis } from '@/services/crisisEscalationService';
 import { useAuth } from '@/contexts/AuthContext';
 import { assessmentToObservation } from '@/fhir/convertAssessment';
@@ -27,7 +27,7 @@ interface Props {
 
 const PHQ2Assessment: React.FC<Props> = ({ onComplete }) => {
   const [responses, setResponses] = useState<number[]>(Array(2).fill(-1));
-  const { log } = useAuditLogger();
+  const { log } = useSecureAuditLogger();
   const { user } = useAuth();
 
   const handleSelect = (index: number, value: number) => {
