+++ src/hooks/useSecureAuditLogger.ts
@@ -1,7 +1,9 @@
 
 import { useAuth } from '@/contexts/AuthContext';
-import { secureServerLogEvent } from '@/services/secureServerAuditLogService';
-import { formRateLimiter } from '@/lib/inputValidation';
+import { EnhancedSecurityAuditService } from '@/services/enhancedSecurityAuditService';
+import { formRateLimiter } from '@/lib/enhancedInputValidation';
+// DEDUPLICATION: Replaces useAuditLogger and useServerSideAuditLogger
+// Reason: provides RLS-compliant logging with rate limiting
 
 /**
  * Hook for secure audit logging using server-side encryption only
@@ -24,28 +26,25 @@ export const useSecureAuditLogger = () => {
       return;
     }
 
-    await secureServerLogEvent({ 
-      action, 
-      details, 
-      userId: user.id 
+    await EnhancedSecurityAuditService.logSecurityEvent({
+      action,
+      details,
     });
   };
 
   const logSecurityEvent = async (eventType: string, details?: Record<string, any>) => {
-    await log(`SECURITY_${eventType}`, {
-      event_type: eventType,
-      timestamp: new Date().toISOString(),
-      ...details
+    await EnhancedSecurityAuditService.logSecurityEvent({
+      action: eventType,
+      details: {
+        event_type: eventType,
+        timestamp: new Date().toISOString(),
+        ...details,
+      },
     });
   };
 
   const logDataAccess = async (table: string, operation: string, recordCount: number = 1) => {
-    await log('DATA_ACCESS', {
-      table,
-      operation,
-      record_count: recordCount,
-      timestamp: new Date().toISOString()
-    });
+    await EnhancedSecurityAuditService.logDataAccessEvent(table, operation, recordCount);
   };
   
   return {
