+++ src/services/enhancedRealtimeService.ts
@@ -2,6 +2,13 @@
 import { RealtimeChannel } from '@supabase/supabase-js';
 import { supabase } from '@/integrations/supabase/client';
 import { debugService } from './debugService';
+import { RealtimeAlert, RealtimePresence } from './realtime/types';
+
+// DEDUPLICATION: Candidate to replace `realtimeService.ts`
+// Reason: provides connection health monitoring and reconnection logic.
+// Missing features: alert broadcasting and presence updates currently in
+// `realtimeService.ts`. Consumers using those features must be updated
+// once functionality is merged.
 
 interface ConnectionHealth {
   isConnected: boolean;
@@ -17,6 +24,11 @@ export class EnhancedRealtimeService {
   private lastPing = Date.now();
   private healthCheckInterval: NodeJS.Timeout | null = null;
   private reconnectTimeout: NodeJS.Timeout | null = null;
+  private presenceChannel: RealtimeChannel | null = null;
+  private alertHandlers = new Map<string, (alert: RealtimeAlert) => void>();
+  private presenceHandlers = new Set<(presence: RealtimePresence[]) => void>();
+  private userId: string | null = null;
+  private isInitialized = false;
   
   constructor() {
     this.startHealthMonitoring();
@@ -86,12 +98,203 @@ export class EnhancedRealtimeService {
       
       return channel;
     } catch (error) {
-      debugService.log('error', 'Failed to subscribe to channel', { 
-        channelName, 
-        error: error.message 
+      debugService.log('error', 'Failed to subscribe to channel', {
+        channelName,
+        error: error.message
       });
       throw error;
     }
+  }
+
+  async initialize(userId: string): Promise<void> {
+    if (this.isInitialized && this.userId === userId) return;
+
+    await this.cleanup();
+
+    this.userId = userId;
+
+    await this.subscribeAlertChannel(userId);
+    await this.subscribePresenceChannel();
+
+    this.isInitialized = true;
+    debugService.log('realtime', 'Enhanced realtime initialized', { userId });
+  }
+
+  private async subscribeAlertChannel(userId: string): Promise<void> {
+    const channelName = `alerts:${userId}`;
+    const channel = await this.subscribe(channelName);
+    channel.on('broadcast', { event: 'alert' }, (payload) => {
+      this.handleAlert(payload.payload as RealtimeAlert);
+    });
+  }
+
+  private async subscribePresenceChannel(): Promise<void> {
+    const channel = await this.subscribe('presence');
+    channel
+      .on('presence', { event: 'sync' }, () => {
+        const state = channel.presenceState();
+        this.handlePresenceUpdate(this.transformPresenceState(state));
+      })
+      .on('presence', { event: 'join' }, () => {
+        const state = channel.presenceState();
+        this.handlePresenceUpdate(this.transformPresenceState(state));
+      })
+      .on('presence', { event: 'leave' }, () => {
+        const state = channel.presenceState();
+        this.handlePresenceUpdate(this.transformPresenceState(state));
+      });
+
+    this.presenceChannel = channel;
+  }
+
+  private transformPresenceState(state: any): RealtimePresence[] {
+    const presence: RealtimePresence[] = [];
+    Object.keys(state).forEach((key) => {
+      const userPresence = state[key];
+      if (userPresence && userPresence.length > 0) {
+        const latest = userPresence[0];
+        presence.push({
+          userId: latest.userId || key,
+          userName: latest.userName || 'Anonymous',
+          status: latest.status || 'online',
+          lastSeen: latest.lastSeen || new Date().toISOString()
+        });
+      }
+    });
+    return presence;
+  }
+
+  async sendAlert(recipientIds: string[], alert: Omit<RealtimeAlert, 'id' | 'timestamp'>): Promise<void> {
+    const { data: { user } } = await supabase.auth.getUser();
+    if (!user) throw new Error('User not authenticated');
+
+    const fullAlert: RealtimeAlert = {
+      ...alert,
+      id: crypto.randomUUID(),
+      timestamp: new Date().toISOString()
+    };
+
+    await Promise.all(
+      recipientIds.map((id) =>
+        supabase.channel(`alerts:${id}`).send({ type: 'broadcast', event: 'alert', payload: fullAlert })
+      )
+    );
+
+    this.lastPing = Date.now();
+  }
+
+  async sendCrisisAlert(message: string, location?: string): Promise<void> {
+    const { data: { user } } = await supabase.auth.getUser();
+    if (!user) throw new Error('User not authenticated');
+
+    const [profileResult, contactsResult] = await Promise.all([
+      supabase.from('profiles').select('full_name').eq('id', user.id).single(),
+      supabase.from('support_contacts').select('id').eq('user_id', user.id)
+    ]);
+
+    if (profileResult.error || contactsResult.error) {
+      throw new Error('Failed to fetch user data');
+    }
+
+    const profile = profileResult.data;
+    const contacts = contactsResult.data || [];
+
+    const recipientIds = contacts.map((c) => c.id);
+
+    await this.sendAlert(recipientIds, {
+      type: 'crisis',
+      senderId: user.id,
+      senderName: profile?.full_name || 'Unknown',
+      message,
+      urgency: 'high',
+      location
+    });
+
+    if (this.presenceChannel) {
+      await this.presenceChannel.track({
+        userId: user.id,
+        userName: profile?.full_name || 'Anonymous',
+        status: 'in-crisis',
+        lastSeen: new Date().toISOString()
+      });
+    }
+
+    this.lastPing = Date.now();
+  }
+
+  async updateStatus(status: 'online' | 'away' | 'in-crisis'): Promise<void> {
+    const { data: { user } } = await supabase.auth.getUser();
+    if (!user || !this.presenceChannel) return;
+
+    const { data: profile } = await supabase.from('profiles').select('full_name').eq('id', user.id).single();
+    await this.presenceChannel.track({
+      userId: user.id,
+      userName: profile?.full_name || 'Anonymous',
+      status,
+      lastSeen: new Date().toISOString()
+    });
+    this.lastPing = Date.now();
+  }
+
+  onAlert(handler: (alert: RealtimeAlert) => void): () => void {
+    const id = crypto.randomUUID();
+    this.alertHandlers.set(id, handler);
+    return () => {
+      this.alertHandlers.delete(id);
+    };
+  }
+
+  onPresenceUpdate(handler: (presence: RealtimePresence[]) => void): () => void {
+    this.presenceHandlers.add(handler);
+    return () => {
+      this.presenceHandlers.delete(handler);
+    };
+  }
+
+  getDebugInfo() {
+    return {
+      channelCount: this.channels.size,
+      lastPing: this.lastPing,
+      reconnectAttempts: this.reconnectAttempts
+    };
+  }
+
+  async cleanup(): Promise<void> {
+    this.unsubscribeAll();
+    if (this.presenceChannel) {
+      try {
+        await this.presenceChannel.untrack();
+      } catch {
+        // ignore
+      }
+      this.presenceChannel = null;
+    }
+    this.alertHandlers.clear();
+    this.presenceHandlers.clear();
+    this.isInitialized = false;
+    this.userId = null;
+  }
+
+  private handleAlert(alert: RealtimeAlert): void {
+    this.lastPing = Date.now();
+    this.alertHandlers.forEach((handler) => {
+      try {
+        handler(alert);
+      } catch (error) {
+        debugService.log('error', 'Alert handler error', { error });
+      }
+    });
+  }
+
+  private handlePresenceUpdate(presence: RealtimePresence[]): void {
+    this.lastPing = Date.now();
+    this.presenceHandlers.forEach((handler) => {
+      try {
+        handler(presence);
+      } catch (error) {
+        debugService.log('error', 'Presence handler error', { error });
+      }
+    });
   }
 
   private async handleDisconnect(channelName?: string): Promise<void> {
@@ -234,5 +437,73 @@ export class EnhancedRealtimeService {
     }
   }
 }
+
+export const subscribeToCrisisEvents = (userId: string, callback: (payload: any) => void): RealtimeChannel => {
+  return supabase
+    .channel('crisis_events')
+    .on(
+      'postgres_changes',
+      {
+        event: 'INSERT',
+        schema: 'public',
+        table: 'crisis_events',
+        filter: `user_id=eq.${userId}`
+      },
+      callback
+    )
+    .subscribe();
+};
+
+export const subscribeToMoodUpdates = (userId: string, callback: (payload: any) => void): RealtimeChannel => {
+  return supabase
+    .channel('daily_checkins')
+    .on(
+      'postgres_changes',
+      {
+        event: 'UPDATE',
+        schema: 'public',
+        table: 'daily_checkins',
+        filter: `user_id=eq.${userId}`
+      },
+      callback
+    )
+    .subscribe();
+};
+
+export const unsubscribeFromChannel = (channel: RealtimeChannel): void => {
+  supabase.removeChannel(channel);
+};
+
+export const subscribeToAllCheckInUpdates = (userId: string, callback: (payload: any) => void): RealtimeChannel => {
+  return supabase
+    .channel('all_checkins')
+    .on(
+      'postgres_changes',
+      {
+        event: '*',
+        schema: 'public',
+        table: 'daily_checkins',
+        filter: `user_id=eq.${userId}`
+      },
+      callback
+    )
+    .subscribe();
+};
+
+export const subscribeToEmergencyContactUpdates = (userId: string, callback: (payload: any) => void): RealtimeChannel => {
+  return supabase
+    .channel('emergency_contacts')
+    .on(
+      'postgres_changes',
+      {
+        event: '*',
+        schema: 'public',
+        table: 'emergency_contacts',
+        filter: `user_id=eq.${userId}`
+      },
+      callback
+    )
+    .subscribe();
+};
 
 export const enhancedRealtimeService = new EnhancedRealtimeService();
